Move python script to other scripts
Add GNUmakefile as cmake wrapper to simplify the development
